【更新】自定义字符串拷贝，避免 snprintf 一个字节的时候，对源字符串截断 【更新】支持关闭颜色字体 【更新】hexdump 支持任意宽度

+++ inc/mlog.h
@@ -25,7 +25,7 @@ extern "C"{
 
 /* init */
 int mlog_init(void);
-void mlog_output(uint32_t level, const char *tag, const char *format, ...);
+void mlog_output(uint8_t level, const char *tag, const char *format, ...);
 void mlog_raw(const char *format, ...);
 void mlog_hexdump(const char *tag, uint8_t *buf, size_t len);
 /* async */
@@ -37,7 +37,7 @@ int mlog_backend_register(mlog_backend_t *backend, const char *name, int sup_col
 int mlog_backend_unregister(mlog_backend_t *backend);
 mlog_backend_t *mlog_backend_find(const char *name);
 /* filter */
-void mlog_global_filter_level_set(uint32_t level);
+void mlog_global_filter_level_set(uint8_t level);
 void mlog_global_filter_tag_set(const char *tag);
 
 /* port */
